+++ raphael.svg.js
@@ -696,6 +696,7 @@ window.Raphael.svg && function (R) {
      * Element.rotate
      [ method ]
      **
+     * Deprecated! Use @Element.transform instead.
      * Adds rotation by given angle around given point to the list of
      * transformations of the element.
      > Parameters
@@ -728,6 +729,7 @@ window.Raphael.svg && function (R) {
      * Element.scale
      [ method ]
      **
+     * Deprecated! Use @Element.transform instead.
      * Adds scale by given amount relative to given point to the list of
      * transformations of the element.
      > Parameters
@@ -763,6 +765,7 @@ window.Raphael.svg && function (R) {
      * Element.translate
      [ method ]
      **
+     * Deprecated! Use @Element.transform instead.
      * Adds translation by given amount to the list of transformations of the element.
      > Parameters
      - dx (number) horisontal shift
@@ -866,12 +869,12 @@ window.Raphael.svg && function (R) {
      * Removes element form the paper.
     \*/
     elproto.remove = function () {
-        if (this.removed) {
+        if (this.removed || !this.node.parentNode) {
             return;
         }
         var paper = this.paper;
         paper.__set__ && paper.__set__.exclude(this);
-        eve.unbind("*.*." + this.id);
+        eve.unbind("raphael.*.*." + this.id);
         if (this.gradient) {
             paper.defs.removeChild(this.gradient);
         }
@@ -1031,7 +1034,7 @@ window.Raphael.svg && function (R) {
             params = name;
         }
         for (var key in params) {
-            eve("attr." + key + "." + this.id, this, params[key]);
+            eve("raphael.attr." + key + "." + this.id, this, params[key]);
         }
         for (key in this.paper.customAttributes) if (this.paper.customAttributes[has](key) && params[has](key) && R.is(this.paper.customAttributes[key], "function")) {
             var par = this.paper.customAttributes[key].apply(this, [].concat(params[key]));
@@ -1252,7 +1255,7 @@ window.Raphael.svg && function (R) {
         return container;
     };
     R._engine.setViewBox = function (x, y, w, h, fit) {
-        eve("setViewBox", this, this._viewBox, [x, y, w, h, fit]);
+        eve("raphael.setViewBox", this, this._viewBox, [x, y, w, h, fit]);
         var size = mmax(w / this.width, h / this.height),
             top = this.top,
             aspectRatio = fit ? "meet" : "xMinYMin",
@@ -1321,7 +1324,7 @@ window.Raphael.svg && function (R) {
      * Clears the paper, i.e. removes all the elements.
     \*/
     R.prototype.clear = function () {
-        R.eve("clear", this);
+        R.eve("raphael.clear", this);
         var c = this.canvas;
         while (c.firstChild) {
             c.removeChild(c.firstChild);
@@ -1338,7 +1341,7 @@ window.Raphael.svg && function (R) {
      * Removes the paper from the DOM.
     \*/
     R.prototype.remove = function () {
-        eve("remove", this);
+        eve("raphael.remove", this);
         this.canvas.parentNode && this.canvas.parentNode.removeChild(this.canvas);
         for (var i in this) {
             this[i] = typeof this[i] == "function" ? R._removedFactory(i) : null;