+++ history.md
@@ -1,4 +1,24 @@
-2.0.2 * 2012-02-08
+2.1.0 • 2012-03-14
+------------------
+
+  * Put all Eve events into “raphael.” namespace
+  * Refactored path caching
+  * Added x2 and y2 to returning bounding box values
+  * Fixed bug with matrix.f in animation
+  * Method [Paper.print](http://raphaeljs.com/reference.html#Paper.print) now returns all letters as one path without any transformations applied and supports multiline text
+  * New methods
+    * [Element.isPointInside](http://raphaeljs.com/reference.html#Element.isPointInside)
+    * [Paper.getElementsByPoint](http://raphaeljs.com/reference.html#Paper.getElementsByPoint)
+    * [Raphael.bezierBBox](http://raphaeljs.com/reference.html#Raphael.bezierBBox)
+    * [Raphael.isBBoxIntersect](http://raphaeljs.com/reference.html#Raphael.isBBoxIntersect)
+    * [Raphael.isPointInsideBBox](http://raphaeljs.com/reference.html#Raphael.isPointInsideBBox)
+    * [Raphael.isPointInsidePath](http://raphaeljs.com/reference.html#Raphael.isPointInsidePath)
+    * [Raphael.pathBBox](http://raphaeljs.com/reference.html#Raphael.pathBBox)
+    * [Raphael.pathIntersection](http://raphaeljs.com/reference.html#Raphael.pathIntersection)
+    * [Raphael.toMatrix](http://raphaeljs.com/reference.html#Raphael.toMatrix)
+    * [Raphael.transformPath](http://raphaeljs.com/reference.html#Raphael.transformPath)
+
+2.0.2 • 2012-02-08
 ------------------
 
   * Removing of linked element now removes `<a>` as well
@@ -10,7 +30,7 @@
   * Fixed bug with changing fill from image to solid colour
   * fixed renderfix method
 
-2.0.1 * 2011-11-18
+2.0.1 • 2011-11-18
 ------------------
 
   * Global variables leakage fix
@@ -25,7 +45,7 @@
   * Fix for `clip-path`
   * Doc update
 
-2.0.0 * 2011-10-03
+2.0.0 • 2011-10-03
 ------------------
 
   * Completely changed transformation handling: